Исправить процесс согласования/утверждения ППРа, составления рапортов Fixes #85
